third_party/blink/web_tests/editing/execCommand/indent/indent_pre_with_new_lines.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>
<div id="log"></div>
<script>
test(() => {
  assert_selection(
    '<div contenteditable style="white-space: pre">^\nHello|</div>',
    'indent',
    '<div contenteditable style="white-space: pre"><blockquote style="margin: 0 0 0 40px; border: none; padding: 0px;">^\nHello|</blockquote></div>');
}, 'Indent a block that preserves white spaces. The block has texts started with a new line.');
test(() => {
  assert_selection(
    '<div contenteditable style="white-space: pre">^\n<div>Hello|</div></div>',
    'indent',
    '<div contenteditable style="white-space: pre"><blockquote style="margin: 0 0 0 40px; border: none; padding: 0px;">^\n<div>Hello|</div></blockquote></div>');
}, 'Indent a block that preserves white spaces. The block has a new line text and a div.');
</script>
